move EncryptColumnToken to rewrite package
